requirements.txt
django
djangorestframework
drf_yasg
psycopg2-binary>=2.8